+++ src/ast.h
@@ -438,7 +438,6 @@ struct ast_s {
         } LangDef;
         struct {
             ast_t *indexed, *index;
-            bool unchecked;
         } Index;
         struct {
             ast_t *fielded;