+++ src/compile/optionals.c
@@ -8,6 +8,7 @@
 #include "../typecheck.h"
 #include "../types.h"
 #include "compilation.h"
+#include "indexing.h"
 
 Text_t optional_into_nonnone(type_t *t, Text_t value) {
     if (t->tag == OptionalType) t = Match(t, OptionalType)->type;
@@ -122,6 +123,7 @@ Text_t compile_optional(env_t *env, ast_t *ast) {
 public
 Text_t compile_non_optional(env_t *env, ast_t *ast) {
     ast_t *value = Match(ast, NonOptional)->value;
+    if (value->tag == Index && Match(value, Index)->index != NULL) return compile_indexing(env, value, true);
     type_t *t = get_type(env, value);
     Text_t value_code = compile(env, value);
     int64_t line = get_line_number(ast->file, ast->start);
@@ -129,5 +131,5 @@ Text_t compile_non_optional(env_t *env, ast_t *ast) {
         "({ ", compile_declaration(t, Text("opt")), " = ", value_code, "; ", "if unlikely (",
         check_none(t, Text("opt")), ")\n", "#line ", line, "\n", "fail_source(", quoted_str(ast->file->filename), ", ",
         (int64_t)(value->start - value->file->text), ", ", (int64_t)(value->end - value->file->text), ", ",
-        "\"This was expected to be a value, but it's none\");\n", optional_into_nonnone(t, Text("opt")), "; })");
+        "\"This was expected to be a value, but it's `none`\\n\");\n", optional_into_nonnone(t, Text("opt")), "; })");
 }